fix: blockentity and entity

+++ src/main/java/cn/nukkit/event/inventory/FurnaceBurnEvent.java
@@ -17,10 +17,10 @@ public class FurnaceBurnEvent extends BlockEvent implements Cancellable {
 
     private final BlockEntityFurnace furnace;
     private final Item fuel;
-    private short burnTime;
+    private int burnTime;
     private boolean burning = true;
 
-    public FurnaceBurnEvent(BlockEntityFurnace furnace, Item fuel, short burnTime) {
+    public FurnaceBurnEvent(BlockEntityFurnace furnace, Item fuel, int burnTime) {
         super(furnace.getBlock());
         this.fuel = fuel;
         this.burnTime = burnTime;
@@ -35,11 +35,11 @@ public class FurnaceBurnEvent extends BlockEvent implements Cancellable {
         return fuel;
     }
 
-    public short getBurnTime() {
+    public int getBurnTime() {
         return burnTime;
     }
 
-    public void setBurnTime(short burnTime) {
+    public void setBurnTime(int burnTime) {
         this.burnTime = burnTime;
     }
